refactor: 토큰 관련 설정 패키지 변경
