+++ src/main/java/com/management/tab/infrastructure/jwt/JwtDecoder.java
@@ -2,7 +2,7 @@ package com.management.tab.infrastructure.jwt;
 
 import com.management.tab.domain.auth.PrivateClaims;
 import com.management.tab.domain.auth.TokenDecoder;
-import com.management.tab.config.auth.security.config.properties.TokenProperties;
+import com.management.tab.config.properties.TokenProperties;
 import com.management.tab.config.auth.security.enums.TokenType;
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.JWEDecrypter;
