+++ src/test/java/com/management/tab/infrastructure/jwt/JwtEncoderTest.java
@@ -2,7 +2,7 @@ package com.management.tab.infrastructure.jwt;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.management.tab.config.auth.security.config.properties.TokenProperties;
+import com.management.tab.config.properties.TokenProperties;
 import com.management.tab.config.auth.security.enums.TokenType;
 import com.nimbusds.jose.JWEEncrypter;
 import com.nimbusds.jose.KeyLengthException;
